# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "EdgeCraft Swagger API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/codes": {
            "get": {
                "description": "Get all code list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code"
                ],
                "summary": "GetCodeList",
                "operationId": "GetCodeList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "put": {
                "description": "Update code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code"
                ],
                "summary": "UpdateCode",
                "operationId": "UpdateCode",
                "parameters": [
                    {
                        "description": "Code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Code"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "Register code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code"
                ],
                "summary": "SetCode",
                "operationId": "SetCode",
                "parameters": [
                    {
                        "description": "Code",
                        "name": "code",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Code"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/codes/groups": {
            "get": {
                "description": "Get all code-group list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CodeGroup"
                ],
                "summary": "GetCodeGroupList",
                "operationId": "GetCodeGroupList",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "put": {
                "description": "Update code group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CodeGroup"
                ],
                "summary": "UpdateCodeGroup",
                "operationId": "UpdateCodeGroup",
                "parameters": [
                    {
                        "description": "Code Group",
                        "name": "codeGroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CodeGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "post": {
                "description": "Register code group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CodeGroup"
                ],
                "summary": "SetCodeGroup",
                "operationId": "SetCodeGroup",
                "parameters": [
                    {
                        "description": "Code Group",
                        "name": "codeGroup",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CodeGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/codes/groups/{groupId}": {
            "get": {
                "description": "Get code group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CodeGroup"
                ],
                "summary": "GetCodeGroup",
                "operationId": "GetCodeGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete code group and codes belong to",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CodeGroup"
                ],
                "summary": "DeleteCodeGroup",
                "operationId": "DeleteCodeGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/codes/{groupId}": {
            "get": {
                "description": "Get codes by group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code"
                ],
                "summary": "GetCodeListByGroup",
                "operationId": "GetCodeListByGroup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
        "/codes/{groupId}/{code}": {
            "get": {
                "description": "Get code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code"
                ],
                "summary": "GetCode",
                "operationId": "GetCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete code",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Code"
                ],
                "summary": "DeleteCode",
                "operationId": "DeleteCode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Code Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Code",
                        "name": "code",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ReturnData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Code": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1
                },
                "desc": {
                    "type": "string",
                    "example": "Test Code for Testing"
                },
                "display_order": {
                    "type": "integer",
                    "example": 0
                },
                "group_id": {
                    "type": "string",
                    "example": "TestGroup"
                },
                "name": {
                    "type": "string",
                    "example": "TestCode #1"
                },
                "use_yn": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "model.CodeGroup": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string",
                    "example": "Code Group Testing"
                },
                "group_id": {
                    "type": "string",
                    "example": "TestGroup"
                },
                "use_yn": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "response.ReturnData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "isError": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}